correções no crud e documentação

+++ src/components/ButtonDelete/index.jsx
@@ -3,24 +3,27 @@ import axios from 'axios';
 import Button from 'react-bootstrap/Button';
 import Modal from 'react-bootstrap/Modal';
 
-function ButtonDelete({ itemId }) {
+function ButtonDelete({ itemId, onDelete }) {  // Adicionando onDelete como prop
   const [show, setShow] = useState(false);
   const [erro, setErro] = useState(null);
 
   const handleClose = () => setShow(false);
   const handleShow = () => setShow(true);
 
-  const handleDelete = () => {
-    axios.delete(`https://backend-julho-2024.onrender.com/item/${itemId}`)
-      .then(res => {
-        console.log('Item deletado com sucesso:', res.data);
+  const handleDelete = async () => {
+    if (itemId && itemId.length === 24) {
+      try {
+        await axios.delete(`https://backend-julho-2024.onrender.com/item/${itemId}`);
+        onDelete(itemId);  // Chamando a função de exclusão passada como prop
         handleClose();
-        // Opcional: Chamar uma função para atualizar a lista de itens na página
-      })
-      .catch(error => {
-        console.error('Erro ao deletar o item:', error);
-        setErro('Erro ao deletar o item.');
-      });
+      } catch (error) {
+        console.error('Erro ao deletar item:', error);
+        setErro('Erro ao deletar item.');
+      }
+    } else {
+      console.error('ID inválido:', itemId);
+      setErro('ID inválido.');
+    }
   };
 
   return (
@@ -36,11 +39,11 @@ function ButtonDelete({ itemId }) {
         keyboard={false}
       >
         <Modal.Header closeButton>
-          <Modal.Title>Confirmar Deleção</Modal.Title>
+          <Modal.Title>Deletar item</Modal.Title>
         </Modal.Header>
         <Modal.Body>
           {erro && <p className="text-danger">{erro}</p>}
-          <p>Tem certeza que deseja deletar este item?</p>
+          <p>Tem certeza de que deseja deletar este item?</p>
         </Modal.Body>
         <Modal.Footer>
           <Button variant="secondary" onClick={handleClose}>
